Add a flatten option to move handler to the top of the zip
In a mono-repo structure we can have lambda functions nested within
folders. Keeping track in which folder they are and having to update
the handler name inside of aws can get tedious and prone to errors.

The flatten option will, given the name of the function when called:
lambpack functions/function-one/index.js function.zip

replace the path leading up to the function in all dependencies so they
appear inside the zip as:
function.zip
  - index.js
  - index-dep.js

The flatten mechanism checks if a relative dependency require call would
break once applied and abort with an error message explaining which
dependency prevents flatten to work and a recommendation to disable it
if needed.